requirements.txt
langchain
python-dotenv
langchain-openai
langchain-groq
langchain_google_genai
langchain-community
langchain-core
pypdf
faiss-cpu
structlog
PyMuPDF
pandas
streamlit
langchain-core[tracing]
pytest


-e .